requirements.txt
pika == 1.3.1
pika-stubs == 0.1.3
python-dotenv == 1.0.0
sqlalchemy == 2.0.8
sqlalchemy_utils == 0.40.0
geoip2.database == 4.6.0
sqlalchemy.orm == 2.0.9
sqlalchemy.ext.declarative == 0.1.3
datetime == 5.1
hashlib == 20081119
re
signal
sys
os
urllib.parse
json
geoip2.database
requests == 2.28.2
urllib.request == 1.26.15
python-dateutil==2.8.2

    